Update Btree.java
I have a few items in here that I used for testing and debugging, specifically the toString and print methods. The tree is functioning right now, however the only code I tested it against was this:

for(int i = 0; i<64; i++){
	btree.insert(i);
	System.out.println("   Insert: " + (i+1));
	btree.printTreeLevelOrder();
	System.out.println();
}

I have only tested with sequence length 3 and degree sizes of 2 and 7, but I wanted to get this up here so you can check it out and run any test you want to as well.